Make sharder's scheme reusable
